Revert "Added External Java Resolver."
This change breaks tests and changes the CLI behavior.

There should be no "external Java source". Situations like this should be handled by configuring your classpath or classpath loader instead.

This reverts commit 6b1d7237d9690713ec7385a038f55b2c5700846d and 6b1d7237d9690713ec7385a038f55b2c5700846d.
